Seeds and Replays!
Added seed generation to the hill generator. Check the use seed box and
then set it to whatever int your heart desires!

Fixed rotation issues in replay system by converting it to use
Quaternions instead of EulerAngles. Since we don’t have to make any
manual adjustments to the rotations this is the way to go, the replays
are fantastically smooth rotation wise now.

Added replay car to the main scene and threw your replay tester script
in (on the ReplayHandler object) just to mess with it. Its super fun!